+++ lib/src/internal/table_layout_render_box.dart
@@ -28,6 +28,7 @@ class TableLayoutRenderBox<ROW> extends RenderBox
   RenderBox? _bottomCorner;
 
   EasyTableThemeData _theme;
+
   set theme(EasyTableThemeData value) {
     _theme = value;
   }